requirements.txt
requests
python-dotenv
playwright
PlaywrightCapture==1.27.4
flask